fix bugs

+++ src/main/java/ph/com/paraiso/dao/UserDao.java
@@ -13,8 +13,8 @@ public interface UserDao {
 
 	public User addUser(User user);
 	
-	public User getUserById(Long userid);
+	public User getUserById(Integer userid);
 	public User updateUser(User user);
 	
-	public void deleteById(Long userid);
+	public void deleteById(Integer userid);
 }
